+++ source/material/material.h
@@ -2,9 +2,9 @@
 
 #include "maths.h"
 
-float fresnel_dielectric(float incident_cos_theta, float incident_ior, float transmitted_ior);
+double fresnel_dielectric(double incident_cos_theta, double incident_ior, double transmitted_ior);
 float3 faceforward(const float3& n, const float3& v);
-bool refract(const float3& incident, const float3& normal, float ior, float3& transmitted);
+bool refract(const float3& incident, const float3& normal, double ior, float3& transmitted);
 
 class Material {
 public:
