+++ source/material/specular_reflective_material.h
@@ -4,12 +4,12 @@
 
 class SpecularReflectiveMaterial : public Material {
 public:
-    SpecularReflectiveMaterial(float ior);
+    SpecularReflectiveMaterial(double ior);
 
     float3 bsdf(float3& ingoing, const float3& outgoing, const float2& random) const override;
 
     bool is_specular() const override;
 
 private:
-    float m_ior;
+    double m_ior;
 };
